sahicheck_backend/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pymongo>=3.12.0
psycopg2-binary>=2.9.0
pydantic>=1.8.0
